Move LabelExpression.AutoCompleteSeeder to top level as LabelAutoCompleteSeeder.
This also adds missing nullability annotations.
